fix: env, navigation and build
