+++ .env
@@ -5,6 +5,6 @@ BUILD_PATH=dist/
 INLINE_RUNTIME_CHUNK=false
 REACT_APP_API_HOST=http://localhost:4502
 REACT_APP_AEM_AUTHORIZATION_HEADER='Basic YWRtaW46YWRtaW4='
-REACT_APP_ROOT=/content/aem-react-spa/language-masters/en/home.html
-REACT_APP_PAGE_MODEL_PATH=/content/aem-react-spa/language-masters/en/home.model.json
+REACT_APP_ROOT=/content/aem-react-spa/us/en/home.html
+REACT_APP_PAGE_MODEL_PATH=/content/aem-react-spa/us/en.model.json
 REACT_APP_PROXY_ENABLED=true
